backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
flask-dance==7.0.0

psycopg2-binary==2.9.9

gunicorn==21.2.0

python-dotenv==1.0.1
requests==2.32.3

Werkzeug==3.1.3
itsdangerous==2.2.0